pyproject.toml
[tool.poetry]
name = "rexize"
version = "0.2.4"
description = "CLI to bulk resize and convert images recursively"
authors = ["Joseph Lee <joel@joeworks.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/joelee/rexize"
packages = [{ include = "src" }]
include = ["pyproject.toml", "README.md", "LICENSE"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Topic :: Multimedia :: Graphics",
]
keywords = ["cli", "image", "resize", "convert", "bulk"]

[tool.poetry.scripts]
rexize = "src.main:main"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.3.0"
opencv-python = "^4.9.0.80"
numpy = "^1.26.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
pytest-cov = "^4.1.0"
pre-commit = "^3.6.2"
flake8 = "^7.0.0"
isort = "^5.13.2"
icecream = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]
pythonpath = ["src"]
norecursedirs = [".git", ".github", ".vscode", ".idea", "dist", "__pycache__"]

[tool.coverage.report]
fail_under = 62

[tool.coverage.run]
branch = true
omit = ["test/*"]
include = ["src/*"]
command_line = "-m unittest discover -s tests/"

[tool.black]
line-length = 88
target-version = ["py311"]
include = "(\\.pyi?|\\.ipynb)$"
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.__pycache__
  |\.pytest_cache
  |Dockerfile
)/
'''

[tool.isort]
profile = "black"
py_version = 311
src_paths = ["src", "test"]
line_length = 88
